Update search method and enhance field name extraction
The main search method has been updated to an advanced search method that now accommodates an array as a query. Furthermore, the `getRowNamesByExample` method is improved to exclude field names that begin with 'g_', which are generally used for global fields. A simple search method has also been added to facilitate quick searches using a single string.