Update POST body to send 'fields' data only in Filemaker.js
The code has been modified to focus only on the 'fields' section of the record in the POST request body. This change improves data transfer efficiency and ensures proper deserialization on the server side. Also, unnecessary whitespace in the error handling section of the code has been removed for cleaner code presentation.